ep 9

+++ app/Http/Controllers/ArticlesController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ArticleCreateRequest;
-use App\Http\Requests\ArticleUpdateRequest;
 use App\Models\Article;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Requests\ArticleCreateRequest;
+use App\Http\Requests\ArticleUpdateRequest;
 
 class ArticlesController extends Controller
 {
@@ -14,8 +16,11 @@ class ArticlesController extends Controller
      */
     public function index()
     {
+
+
+
         return view('articles.index', [
-            'articles' => Article::all(),
+            'articles' => Article::visableTo(Auth::user())->get(),
         ]);
     }
 
@@ -24,6 +29,8 @@ class ArticlesController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Article::class);
+
         return view('articles.create');
     }
 
@@ -44,6 +51,13 @@ class ArticlesController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('update', $article);
+
+        // if ( Gate::denies('update', $article)) {
+        //     abort(403);
+        // }
+
+
         return view('articles.edit', [
             'article' => $article
         ]);
@@ -64,6 +78,8 @@ class ArticlesController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', $article);
+
         $article->delete();
 
         return redirect()->route('articles.index');
